fix(manager): resolve React Fast Refresh warnings in AuthContext and SetupContext
Fixes #6

Split context hooks into separate files to enable full Fast Refresh support:
- Create AuthContext.hooks.ts with useAuth hook
- Create SetupContext.hooks.ts with useSetup hook
- Export contexts with eslint-disable comment for hook access
- Update all imports across the codebase

This resolves the react-refresh/only-export-components warnings during
development while maintaining proper hot module replacement.
