+++ manager/src/components/layout/Header.tsx
@@ -15,7 +15,7 @@ import MenuIcon from '@mui/icons-material/Menu'
 import AccountCircleIcon from '@mui/icons-material/AccountCircle'
 import LogoutIcon from '@mui/icons-material/Logout'
 import CodeIcon from '@mui/icons-material/Code'
-import { useAuth } from '../../contexts/AuthContext'
+import { useAuth } from '../../contexts/AuthContext.hooks'
 
 interface HeaderProps {
   onMenuClick: () => void
