+++ manager/src/contexts/AuthContext.hooks.ts
@@ -1,0 +1,10 @@
+import { useContext } from 'react'
+import { AuthContext } from './AuthContext'
+
+export function useAuth() {
+  const context = useContext(AuthContext)
+  if (context === undefined) {
+    throw new Error('useAuth must be used within an AuthProvider')
+  }
+  return context
+}
