+++ manager/src/pages/LoginPage.tsx
@@ -10,7 +10,7 @@ import {
   Alert,
   CircularProgress,
 } from '@mui/material'
-import { useAuth } from '../contexts/AuthContext'
+import { useAuth } from '../contexts/AuthContext.hooks'
 
 export default function LoginPage() {
   const navigate = useNavigate()
